Moved pages/ to src/, Added AdvocateSearch component for search page
