+++ src/app/components/AdvocateSearch/types.ts
@@ -1,0 +1,4 @@
+export interface SpecialtyFilterInput {
+	specialty: string;
+	isChecked: boolean;
+}
